修改api鉴权

+++ app/system/model/enter.go
@@ -18,14 +18,14 @@ type SysUser struct {
 // 解析前端数据
 type SysUserFrontEnd struct {
 	gorm.Model
-	Name     string `gorm:"column:name" json:"name"`
-	Username string `gorm:"column:username" json:"username"`
-	Password string `gorm:"column:password" json:"password"`
-	Phone    string `gorm:"column:phone" json:"phone"`
-	Sex      int64  `gorm:"column:sex" json:"sex"`
-	Birthday string `gorm:"column:birthday" json:"birthday"`
-	Email    string `gorm:"column:email" json:"email"`
-	Roles    []uint `json:"roles"`
+	Name     string  `gorm:"column:name" json:"name"`
+	Username string  `gorm:"column:username" json:"username"`
+	Password string  `gorm:"column:password" json:"password"`
+	Phone    string  `gorm:"column:phone" json:"phone"`
+	Sex      int64   `gorm:"column:sex" json:"sex"`
+	Birthday string  `gorm:"column:birthday" json:"birthday"`
+	Email    string  `gorm:"column:email" json:"email"`
+	Roles    []int64 `json:"roles"`
 }
 
 func (user SysUser) TableName() string {
